Add bubble sort visualizer
The big idea: Suppose you have the bubble sort algorithm written as a
function. With each step in the algorithm, you can print the current
array as it is being sorted. So then, in order to visualize the
algorithm, we simply take this output and make a histogram with
it, with the x-axis being the indices and the y-axies being the random
numbers in the randomized array. As the algorithm is constantly
working at sorting the algorithm, the graph will also constantly be
updating to show the sorting in action.

For visualization, we use chart.js (which is under MIT licensing). It
seems to do the job fine and is relativly simply to use. Alternative
suggestions that are more simple or more faster are encouraged. The
web page runs fine on my computer, but has not been tested on any
others.

On an additional note, I chose to write this in Javascript because
this project seems too small to be written as standalone
operating-system software; it suffices to be used in a web
browser.